create_builddir: use a dynamic number of Make jobs
Instead of using a generic value of number of jobs to pass to make to
build in parallel, use /dev/cpuinfo (faster than the cpuinfo tool) to
find a good value. usually the number of logical cores.
This commit also uses the long "--jobs" option to be more explicit,
and adds the "--load-average" option when using parallel jobs. This
follows the recommendation of the Gentoo MAKEOPTS variable
(https://wiki.gentoo.org/wiki/MAKEOPTS).
Performing this operation in the create_builddir script allows to
compute these values once and for all, but still allows the user to
override the value using "CPUS".

Signed-off-by: Jimmy Durand Wesolowski <jimmy.wesolowski@esrlabs.com>